Add a micro-op modules section since they are so important
The micro-op module section is intended to discuss micro-ops
themselves, in isolation. How other modules use micro-ops should be
discussed in the respective module section, and the overall flow of
micro-ops through the core should be discussed in the micro-op
lifecycle section.

The micro-op lifecycle section is solely for discussing how a micro-op
is created out of an instruction, how it flows, and how it is
eventually retired.